bash: Reduce prompt length
IT gave me user and device names that are too long for use in the
default Bash prompt, especially in combination with tmux panes.
Work around this by hardcoding shorter user and device names.

Note that this is just a quick hack, to be revised in the future.
Also note that this is not portable, but it will do for now.